* commit '8790aa33a1faf199dccdeae4b8dd91aa65cfd821':
  Bugfix in monitoring module
  Deprecated LocalPeer.createContainer method

+++ management/server/subutai-common/src/main/java/org/safehaus/subutai/common/metric/ProcessResourceUsage.java
@@ -7,8 +7,8 @@ package org.safehaus.subutai.common.metric;
 public class ProcessResourceUsage
 {
     private String host;
-    private int usedCpu;
-    private int usedRam;
+    private double usedCpu;
+    private double usedRam;
 
 
     public String getHost()
@@ -17,13 +17,13 @@ public class ProcessResourceUsage
     }
 
 
-    public int getUsedCpu()
+    public double getUsedCpu()
     {
         return usedCpu;
     }
 
 
-    public int getUsedRam()
+    public double getUsedRam()
     {
         return usedRam;
     }
